sprint-2/requirements.txt
confluent-kafka
faust-streaming